+++ json_files/books.json
@@ -1,7 +1,3 @@
 [
-{"title":"sas","author":"asas","id":27}
-,
-{"title":"asssss","author":"ssssssss","id":943}
-,
-{"title":"love","author":"eat","id":561}
+{"title":"love","author":"love","id":237}
 ]
