+++ json_files/rentals.json
@@ -1,5 +1,3 @@
 [
-{"date":"2/2/2000","person":92,"book":27}
-,
-{"date":"8/8/2555","person":184,"book":943}
+{"date":"5/5/2002","person":180,"book":237}
 ]
